refactor: nest GraphQL operations under a rootLoginNotification namespace container
Move the root Query (rootLoginNotificationSettings) + root Mutation
(rootLoginNotificationSaveSettings) flat fields under one hierarchical namespace container,
per the Jahia GraphQL convention.

- Query:    rootLoginNotificationSettings     -> rootLoginNotification { settings }
- Mutation: rootLoginNotificationSaveSettings -> rootLoginNotification { saveSettings }

New holder types RootLoginNotificationQuery/Mutation carry the operations as instance
@GraphQLField methods (the GqlSettings return type moves into the query holder); the
*Extension classes keep a single rootLoginNotification accessor field.
@GraphQLRequiresPermission stays on each leaf. Frontend gql.js + jsx and Cypress
fixtures/specs updated in lockstep.
