+++ src/javascript/RootLoginNotification/RootLoginNotification.gql.js
@@ -2,17 +2,21 @@ import {gql} from '@apollo/client';
 
 export const GET_SETTINGS = gql`
     query {
-        rootLoginNotificationSettings {
-            recipient
-            sender
-            subject
-            body
+        rootLoginNotification {
+            settings {
+                recipient
+                sender
+                subject
+                body
+            }
         }
     }
 `;
 
 export const SAVE_SETTINGS = gql`
     mutation RootLoginNotificationSaveSettings($recipient: String, $sender: String, $subject: String, $body: String) {
-        rootLoginNotificationSaveSettings(recipient: $recipient, sender: $sender, subject: $subject, body: $body)
+        rootLoginNotification {
+            saveSettings(recipient: $recipient, sender: $sender, subject: $subject, body: $body)
+        }
     }
 `;
