+++ tests/cypress/fixtures/graphql/mutation/saveSettings.graphql
@@ -1,8 +1,10 @@
 mutation saveSettings($recipient: String, $sender: String, $subject: String, $body: String) {
-    rootLoginNotificationSaveSettings(
-        recipient: $recipient
-        sender: $sender
-        subject: $subject
-        body: $body
-    )
+    rootLoginNotification {
+        saveSettings(
+            recipient: $recipient
+            sender: $sender
+            subject: $subject
+            body: $body
+        )
+    }
 }
